2.1. Data Persister Decoration
In the last course, we create this UserDataPersister class

Um, what is a "Persister"?
Now, let's back up real quick. Whenever you use a POST or PUT endpoint, after ApiPlatform deserializes the data into an object and validates it, it tries to save or persist that object. Usually, this means that we're saving an entity object to the database via Doctrine. But we could "save" an object anywhere, like by sending the data to another API, or putting it into Redis or ElasticSearch. ApiPlatform doesn't really care. These "things that save the object" are called "data persisters".
So far, our two API resources - CheeseListing and User - are both entities... and ApiPlatform has special support for Doctrine, including a core Doctrine data persister. So whenever we make a POST, PUT or also PATCH request to an ApiResource that is an entity, that core Doctrine data persister jumps into action and saves the object to the database. This means that normally we don't need to create our own data persister: the core one calls persist() and flush() for us.

Taking Action Before/After Save
But what if you need to do something right before or after an item is saved to the database? Well, you could use a Doctrine listener for that... but if you want that code to only run in the context of your API, how can we? The answer is to create a custom data persister. For example, we created UserDataPersister because we needed to encode the plain password and set it onto the password field before saving:

 48 lines  src/DataPersister/UserDataPersister.php
class UserDataPersister implements DataPersisterInterface
{
    public function persist($data)
    {
        if ($data->getPlainPassword()) {
            $data->setPassword(
                $this->userPasswordEncoder->encodePassword($data, $data->getPlainPassword())
            );
            $data->eraseCredentials();
        }
        $this->entityManager->persist($data);
        $this->entityManager->flush();
    }
}
To create this data persister, we implemented DataPersisterInterface and added a supports() method that says that we support User objects:

 48 lines  src/DataPersister/UserDataPersister.php
class UserDataPersister implements DataPersisterInterface
{
    public function supports($data): bool
    {
        return $data instanceof User;
    }
}
As soon as we did that, we became 100% responsible for persisting User objects. What I mean is, the normal, core Doctrine data persister will no longer be called for User objects: our persister takes precedence. This is why we did our custom work up here, but then had to call persist() and flush() at the bottom:

 48 lines  src/DataPersister/UserDataPersister.php
class UserDataPersister implements DataPersisterInterface
{
    public function persist($data)
    {
        if ($data->getPlainPassword()) {
            $data->setPassword(
                $this->userPasswordEncoder->encodePassword($data, $data->getPlainPassword())
            );
            $data->eraseCredentials();
        }
        $this->entityManager->persist($data);
        $this->entityManager->flush();
    }
}
To prove that our data persister replaces the core persister for User objects, let's comment-out the persist() call and run our tests
Open tests/Functional/UserResourceTest. This testCreateUser() method should now fail because the User won't actually be saved to the database:

 75 lines  tests/Functional/UserResourceTest.php
class UserResourceTest extends CustomApiTestCase
{
    public function testCreateUser()
    {
        $client = self::createClient();
        $client->request('POST', '/api/users', [
            'json' => [
                'email' => 'cheeseplease@example.com',
                'username' => 'cheeseplease',
                'password' => 'brie'
            ]
        ]);
        $this->assertResponseStatusCodeSame(201);
        $this->logIn($client, 'cheeseplease@example.com', 'brie');
    }
}
Copy that method name and run:

symfony run bin/phpunit --filter=testCreateUser
And... yea! It failed! This test passed a minute ago, but now we get a 400 bad request... which happens because the un-persisted entity is missing an id... and so ApiPlatform has trouble serializing it.