2.2. Data Persister Decoration
If we put the persist() back
And run the test again:

symfony run bin/phpunit --filter=testCreateUser
All better!

Decorating the Core Persister
So the fact that we are now entirely responsible for persisting the object is... not really that big of a deal. But it would be even better if we could do our custom work up here... and then just call the core data persister so that it can do its normal logic.
And that's totally possible via decoration. Yep, instead of injecting the entity manager into the constructor, replace this with DataPersisterInterface and call it, how about, $decoratedDataPersister. Copy that and rename the variable and the $entityManager property to $decoratedDataPersister. This is nice because, down here, instead persist() and flush(), all we need is $this->decoratedDataPersister->persist($data). Down in remove(), we can do the same: $this->decoratedDataPersister->remove($data)
Beautiful! But... before we try this... um... don't try this unless you have Xdebug installed. Because... when I run my test:

symfony run bin/phpunit --filter=testCreateUser
Ah! It's recursion!

  Maximum function nesting level of 256 reached
Ok, let's figure out what's going on. In the constructor, we just said DataPersisterInterface and Symfony, apparently, figured out what to pass us
This means that the data persister must be passed via autowiring.

Let's go get more info about that autowired service. Run:

php bin/console debug:autowiring Persister
And... here it is! This is an alias for some service called debug.api_platform.data_persister. Ok! Let's find more info about that service:

php bin/console debug:container debug.api_platform.data_persister
Ok: the class name is TraceableChainDataPersister. The key word in the name is "chain". If we dug a few levels deeper, we would find out that the data persister service actually holds a collection of data persisters inside of it!
Basically, whenever something needs to be persisted, ApiPlatform calls persist() on this one data persister. Internally, it then loops over the persisters inside, calls supports() on each one
And then calls persist() on the first persister that supports the object.

So... this is really cool! If you need the entire data persister system, you can autowire this one service and call persist() on it.

The problem is that we are effectively calling ourselves! ApiPlatform originally calls persist on the ChainDataPersister, it calls persist() on us... and then we call persist() back on the ChainDataPersister:

 45 lines  src/DataPersister/UserDataPersister.php
class UserDataPersister implements DataPersisterInterface
{
    public function persist($data)
    {
        return $this->decoratedDataPersister->persist($data);
    }
}
Whoops!

So instead of injecting the entire data persister system, let's inject and use just the one data persister that we know we want: the one that is normally used to save entities. We'll do that next and learn more about how the data persister system is different than some other parts of ApiPlatform, like the context builder system.