4.1. Custom Logic Only for some Operations
Data persisters are the way to do logic before or after something saves. But what if you want to do something only when your object is being created? Or only when it's being updated? Or maybe only when a specific field changes from one value to another? Let's talk about all of that.
Here's our first goal: log a message only when a user is created in our API. And this one is pretty simple. Start by adding a third argument to the constructor for LoggerInterface $logger. I'll hit Alt+Enter and go to "Initialize Properties" as a shortcut to create that property and set it
Down in persist(), since this is an entity, we don't need to do anything fancy to determine if the object is being created versus updated. We can say: if ! $data->getId()
Then log something: $this->logger->info(sprintf()):

  User %s just registered. Eureka!

And pass $data->getEmail()
That's it! Custom code for when a user is created, which in a real app might be sending a registration email or throwing a party!
I'm guessing this will all work... but let's at least make sure we didn't break anything:

  symfony php bin/phpunit --filter=testCreateUser