4.2. Custom Logic Only for some Operations
Custom Code only for a Specific Operation
Cool! So let's do something a bit harder: what if you need to run some custom code but only for a specific operation. In User, we have 2 collection operations and 3 item operations
For the most part, by checking the id, you can pretty much figure out which operation is being used. But if you start also using a PATCH operation or custom operations, then this would not be enough.
Like many parts of ApiPlatform, a data persister has a normal interface - DataPersisterInterface.
But also an optional, stronger interface that gives you access to extra info about what's going on.
Change the interface to ContextAwareDataPersisterInterface.
This extends DataPersisterInterface and the difference now is that all the methods suddenly have an array  argument. Copy that and also add it to persist() and down here on remove()
Beautiful! To see what's inside this array, at the top of persist, dump($context)
And then go run the test again:

symfony php bin/phpunit --filter=testCreateUser
And... there it is! It has resource_class, collection_operation_name and a few other keys. Two things about this. One, this context is not the same as the "serialization context" array that we've talked about a lot. There is some overlap, but mostly this word "context" is just being re-used. Two: the last 3 items relate to ApiPlatform's event system... and probably won't be useful here.

The truly useful item is collection_operation_name, which will be called item_operation_name for an item operation - like a PUT request. I'll run all the user tests:

symfony php bin/phpunit tests/Functional/UserResourceTest.php
And... yep! There's a good example of both situations.