4.3. Custom Logic Only for some Operations
Armed with this info, we are dangerous! Back in the persister, if $context['item_operation_name'] - don't forget your $ - ?? null - to avoid an error when this key does not exist - === 'put'. Then this is the PUT item operation. Log User %s is being updated and pass $data->getId()
I love it! When we run the tests now:

symfony php bin/phpunit tests/Functional/UserResourceTest.php
They still pass! I am assuming that the logs are being written correctly... I would at least test this manually in a real app.

By using the context and the $data->getId(), we have a lot of control to execute custom code in the right situations. But things can get more complex. What if we need to run code only when a specific field changes from one value to another?
For example, our CheeseListing entity has an isPublished field, which - so far - isn't writable in our API at all
Next: let's make it possible for a user to publish their listing... but also execute some code when that happens.