5.2. Publishing a Listing
Making isPublished Writable in the API
Let's fix that! At the top of the @ApiResource annotation, as a reminder, we have a denormalizationContext that sets the serialization groups to cheese:write
So if we want a field to be writeable in our API, it needs that group.
Copy that, scroll down to isPublished, add @Groups({}) and paste. Now, as long as this has a setter - and... yep there is a setIsPublished() method.
It will be writable in the API.

Let's see if it is! Go back to your terminal and run the test again:

  symfony php bin/phpunit --filter=testPublishCheeseListing
And... got it! We can now publish a CheeseListing! But... this was the easy part. The real question is: how can we run custom code only when a CheeseListing is published? So, only when the isPublished field changes from false to true? Let's find out how next.