6.1. Detecting the "Published" State Change
Here's the plan: whenever a CheeseListing becomes published, we need to create a notification that is sent to certain users... maybe some users have subscribed to hear about all the latest new cheeses. To help, sort of, "fake" this system, I've created a CheeseNotification entity:

 53 lines  src/Entity/CheeseNotification.php
class CheeseNotification
{
    /**
     * @ORM\Id
     * @ORM\GeneratedValue
     * @ORM\Column(type="integer")
     */
    private $id;
    /**
     * @ORM\ManyToOne(targetEntity=CheeseListing::class)
     * @ORM\JoinColumn(nullable=false)
     */
    private $cheeseListing;
    /**
     * @ORM\Column(type="string", length=255)
     */
    private $notificationText;
    public function __construct(CheeseListing $cheeseListing, string $notificationText)
    {
        $this->cheeseListing = $cheeseListing;
        $this->notificationText = $notificationText;
    }
}
The goal is to insert a new record into this table each time a CheeseListing is published. It's not a real notification system... but it'll be an easy way for us to figure out how to run custom code in just the right situation.

Asserting the Database has Notifications
Let's start - like usual - with a test! We want to assert that after we send the PUT request, there is exactly one CheeseNotification in the database. Now, thanks to Foundry - the library that helps us insert dummy data both in our data fixtures and in our tests - before each test case, our database is emptied. That will make it very easy to count the number of rows inside the the table. Oh, and if you love this "database resetting" feature of Foundry and want a performance boost in your tests, check out DAMADoctrineTestBundle.
Anyways, thanks to another cool feature from Foundry, doing this count is easy: CheeseNotificationFactory::repository()->assertCount(1).

How nice is that? A dead-simple way to count the records in a table. I'll even give this a nice message in case it fails:

  There should be one notification about being published

I love it! But before we try this... and make sure it fails, let's go one step further. Copy the PUT call from up here and paste it below... because we want to make sure that if we publish the same listing again, it should not create a second notification because... we're not really publishing it again. We're... not doing anything with this second call!
Use the same assert: CheeseNotificationFactory::repository()->assertCount(1)
Ok! Let's try it! Find your terminal and run just this test:

  symfony php bin/phpunit --filter=testPublishCheeseListing
And... failed asserting that 0 is identical to 1:

  There should be one notification about being published.