6.2. Detecting the "Published" State Change
Creating the Data Persister
Ok, so... how can we accomplish this? Well, we know that we need to run some code right before or after the CheeseListing is saved. That means that we need a custom data persister. Cool! We've done that before with UserDataPersister!
Create a new PHP class and call it CheeseListingDataPersister. Poetry. Let's keep this as simple as possible: implement the normal DataPersisterInterface. Go to "Code"->"Generate" - or Command+N on a Mac - select "Implement Methods" and add the three method that we need.
Before we fill in the code, let's immediately inject the doctrine data persister so that we can use it to do the actual saving. Add public function __construct() with DataPersisterInterface $decoratedDataPersister
I'll hit Alt+Enter and use "Initialize properties" to add that property and set it in the method
Down in supports()... wait. I almost forgot! We need to make sure the right data persister is passed to us. Open config/services.yaml and we can just copy the UserDataPersister service and change it to CheeseListingDatePersister, because we need the same doctrine data persister service. Ok, back in the persister, the supports() method is easy: return $data instanceof CheeseListing.
Yep, if a CheeseListing is being saved, we want to handle it. In persist(), call $this->decoratedDataPersister->persist($data). And in remove(), $this->decoratedDataPersister->remove($data)
Congratulations! We just created a data persister that does... the exact same thing as before. We can prove it by running the test:

  symfony php bin/phpunit --filter=testPublishCheeseListing
And enjoying the exact same failure! There are still no notifications.