6.3. Detecting the "Published" State Change
Fetching the Original Data
Back in persist(), the question is: how can we detect if the item was just published? I'll add a little PHPDoc above the method to help my editor: we know that $data will definitely be a CheeseListing object.
So, the easiest thing to do would be to say: if $data->isPublished(), then... the listing was just published! Right?
I'm sure you see the problem: that would tell us that the listing is published... but not that it was necessarily just published... it may have already been published... and the user was just changing the description or something. We simply don't have enough information.
What we really need is access to the original data: the way it looked before it was changed by ApiPlatform. If we had that original data, we could compare that with the current data and determine if the isPublished field did in fact change from false to true. But... how can we get that?
Actually, thanks to Doctrine, this is easy! Doctrine already keeps track of the original data: the data that came back when it originally queried the database for the CheeseListing!
To get the original data, we need the entity manager. Add a second argument to the constructor: EntityManagerInterface $entityManager. Hit Alt+Enter to initialize that new property
Then, in persist(), we can say $originalData = $this->entityManager->getUnitOfWork() - that's a very core object in Doctrine - ->getOriginalEntityData() and pass it $data
How cool is that? Let's dump that $originalData to see what it looks like.
Let's go! Run the test:

  symfony php bin/phpunit --filter=testPublishCheeseListing
And... cool, cool, cool! It's an array where each key represents a field of original data. It shows isPublished was false at the start. We are so dangerous now.