6.4. Detecting the "Published" State Change
Using the Original Data
Remove the dump and add a new variable: $wasAlreadyPublished set to $originalData['isPublished'] ?? false
Because if this is a new entity, $originalData will an empty array. And if this is a new CheeseListing, then it was - of course - not already published.
Finally, if $data->getIsPublished() and not $wasAlreadyPublished, we know that it was - in fact - just published!
So let's create a new notification: $notification = new CheeseNotification() and its constructor accepts the related CheeseListing object and the notification text:

  Cheese listing was created!
Oh, and we need to save this to the database: $this->entityManager->persist($notification) and $this->entityManager->flush()
And yes, you could skip the flush() because, a few lines later, the Doctrine data persister will call that for us. We're... technically saving both the notification and the CheeseListing on this line... but it doesn't really matter.

So... did we do it? Let's find out! At your terminal, go test go!

  symfony php bin/phpunit --filter=testPublishCheeseListing
It passes! Woo!

This "original data" trick is really the key to doing custom things on a simple, RESTful state change, like isPublished going from false to true.
Next: let's use this same trick to accomplish something else. Let's add complex rules around who can publish a CheeseListing and under what conditions. Like... maybe the owner can publish the listing... but maybe only if the description is longer than 100 characters... except that an admin can always publish... even if the description is too short. Yikes! Typical, awesome, confusing business logic. Let's crush it next!