7.1. Validating Who/When Can Publish
I have an idea! Let's complicate things!
We can't let anyone publish a CheeseListing: only the owner or an admin should be able to do that. But we already have that covered thanks to the security attribute on the CheeseListing put operation:

 217 lines  src/Entity/CheeseListing.php
/**
 * @ApiResource(
 *     itemOperations={
 *          "put"={
 *              "security"="is_granted('EDIT', object)",
 *              "security_message"="Only the creator can edit a cheese listing"
 *          },
 *     },
 * )
 */
class CheeseListing
{
}
It has "security"="is_granted('EDIT', object), which activate a custom voter that we created in the last tutorial: CheeseListingVoter:

 55 lines  src/Security/Voter/CheeseListingVoter.php
class CheeseListingVoter extends Voter
{
    protected function supports($attribute, $subject)
    {
        // replace with your own logic
        // https://symfony.com/doc/current/security/voters.html
        return in_array($attribute, ['EDIT'])
            && $subject instanceof CheeseListing;
    }
    protected function voteOnAttribute($attribute, $subject, TokenInterface $token)
    {
        $user = $token->getUser();
        // if the user is anonymous, do not grant access
        if (! $user instanceof UserInterface) {
            return false;
        }
        /** @var CheeseListing $subject */
        // ... (check conditions and return true to grant permission) ...
        switch ($attribute) {
            case 'EDIT':
                if ($subject->getOwner() === $user) {
                    return true;
                }
                if ($this->security->isGranted('ROLE_ADMIN')) {
                    return true;
                }
                return false;
        }
        throw new \Exception(sprintf('Unhandled attribute "%s"', $attribute));
    }
}
This looks for that EDIT attribute and checks to make sure that the current user is the owner of the CheeseListing or is an admin. So it's already true that only the owner or an admin can even hit this put operation.
But now we have a new requirement: to prevent low-quality listings from being added, our business team has told us that an owner can only publish a CheeseListing if the description is at least 100 characters. If it's shorter, publishing should fail with a 400 status code. Oh... but the business people also said that this rule should not apply to admins: they should be able to publish no matter how short the description is.

Using Foundry State to Make a Longer Description
Um... ok! Let's start with a test to describe all this craziness. In CheeseListingResourceTest, to make sure that our testPublishCheeseListing() keeps working, we need to give CheeseListing a longer description.
If you look in CheeseListingFactory - src/Factory/CheeseListingFactory - by default, the description is short:

 58 lines  src/Factory/CheeseListingFactory.php
final class CheeseListingFactory extends ModelFactory
{
    protected function getDefaults(): array
    {
        return [
            'description' => 'What can I say? A raw cube of cheese power',
        ];
    }
}
But I've already added something called a "state method": withLongDescription():

 58 lines  src/Factory/CheeseListingFactory.php
final class CheeseListingFactory extends ModelFactory
{
    public function withLongDescription(): self
    {
        return $this->addState([
            'description' => self::faker()->paragraphs(3, true)
        ]);
    }
}
If we call this on the factory, it will set the description to three paragraphs via Faker... which will definitely be long enough.
In other words, inside the test, we can add ->withLongDescription() to make sure that the CheeseListing created will have a long enough description:

 148 lines  tests/Functional/CheeseListingResourceTest.php
class CheeseListingResourceTest extends CustomApiTestCase
{
    public function testPublishCheeseListing()
    {
        $cheeseListing = CheeseListingFactory::new()
            ->withLongDescription()
            ->create(['owner' => ]);
    }
}
Thanks to this, once we add the new restrictions, this test should still pass.

Testing the Validation Requirements
Below this, I'm going to paste in a new test method, which you can get from the code block on this page
Let's walk through it real quick.

We start by creating a CheeseListing with a short description, logging in as the owner of that listing, trying to publish it and checking for a 400 status code. Next, we log in as an admin user, do the same thing and check that this does work.
I also have a test that logs back in as the normal user and changes something different... just to make sure normal edits still work.
Finally, the last two parts are for unpublishing. In our system, unpublishing is something that only an admin can do. To test that, we log in as a normal user, try to set isPublished to false and make sure it doesn't work.Then we log in as an admin and make sure that it does work.
Phew! And yes, we could - and maybe should - break this test into smaller test methods to make debugging easier. I'm being a bit lazy by combining them.