7.2. Validating Who/When Can Publish
Is this Security? Or Validation?
So how can accomplish all of this? Well first... is this a security check or a validation check? Because sometimes, the difference between the two is blurry.
In theory, we could do something inside the security attribute. The expression has a variable called previous_object, which is the object before it was updated. So, we could compare the previous object's isPublished with the object's isPublished and do different things. Or, to keep this readable, we could even pass both the object and previous_object into a voter as an array. That seems a little crazy to me... but it would probably work!
However, I tend to view things like this: security is best when you're trying to completely prevent access to an operation. Validation is best when the restrictions you need to apply are based on the data that's being sent, like preventing a field from changing to some value. That's true even if those values depend on the authenticated user.

In other words, we're going to create a custom validator! At your terminal, run:

  php bin/console make:validator
I'll call it: ValidIsPublished.
This generated two classes: a class that represents the "annotation" and also a class that holds the actual validator logic. We can see both inside src/Validator/.
Before we add any logic to either class, let's use the annotation in CheeseListing. So, above the CheeseListing class, add @ValidIsPublished()
Now, the reason we're adding this above the class - and not above the  property - is that our validator will need access to the entire CheeseListing object so that it can check both the isPublished and description properties. If we had put this above isPublished, then only that field would be passed to us.
Let's try the test and see what the failure looks like right now. Over in the test class, copy the new method name, and then run:

  symfony php bin/phpunit --filter=testPublishCheeseListingValidation
And... it had an error! Let's see what's going on:

  Constraint ValidIsPublished cannot be put on classes